Separated whole genome controls code into new file
